pyproject.toml
[tool.poetry]
name = "peerlogic-voip"
version = "1.0.0"
description = "Universal VoIP administration system for Peerlogic"
authors = ["Peerlogic <info@peerlogic.com>"]
readme = "README.md"
packages = [{include = "src"}, {include = "config"}]

[tool.poetry.dependencies]
python = "^3.9,<3.12"
Django = ">=4.2,<5.0"
djangorestframework = ">=3.14,<4.0"
django-cors-headers = ">=4.2,<5.0"
django-filter = ">=23.2,<24.0"
psycopg2-binary = ">=2.9,<3.0"
python-dotenv = ">=1.0,<2.0"
httpx = ">=0.24,<1.0"
pydantic = ">=2.0,<3.0"
email-validator = ">=2.0,<3.0"
cryptography = ">=41.0,<42.0"
drf-spectacular = ">=0.26,<1.0"
python-dateutil = ">=2.8,<3.0"
gunicorn = ">=21.0,<22.0"
whitenoise = ">=6.5,<7.0"
dj-database-url = ">=2.1,<3.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4,<8.0"
pytest-django = ">=4.5,<5.0"
pytest-asyncio = ">=0.21,<1.0"
black = ">=23.7,<24.0"
flake8 = ">=6.1,<7.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--reuse-db --nomigrations"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

